feat: purchase-receipt

+++ Proyecto_web/database/migrations/2024_08_23_210125_create_receipts_table.php
@@ -15,7 +15,7 @@ class CreateReceiptsTable extends Migration
     {
         Schema::create('receipts', function (Blueprint $table) {
             $table->id();
-            $table->integer('totalPrice');
+            $table->double('totalPrice');
             $table->date('date');
             $table->timestamps();
         });
